+++ Langauge/Swift/framework/swiftUI/devote/devote/Utility/Constant.swift
@@ -1,0 +1,22 @@
+//
+//  Constant.swift
+//  devote
+//
+//  Created by 유승철 on 1/29/24.
+//
+
+import SwiftUI
+
+// MARK: - Formatter
+let itemFormatter: DateFormatter = {
+    let formatter = DateFormatter()
+    formatter.dateStyle = .short
+    formatter.timeStyle = .medium
+    return formatter
+}()
+
+
+// MARK: - ui
+
+
+// MARK: - ux
